Change ndim return type to usize

+++ src/lib.rs
@@ -164,8 +164,8 @@ impl Tensor {
     /// assert_eq!(Tensor::new(&[1, 2]).ndim(), 2);
     /// assert_eq!(Tensor::new(&[1, 3, 6, 5]).ndim(), 4);
     /// ```
-    pub fn ndim(&self) -> i32 {
-        self.shape.len() as i32
+    pub fn ndim(&self) -> usize {
+        self.shape.len()
     }
 
     /// Transposes the tensor axes according to the permutation.
